Initial cut of theme support.
Decided to go with a simple struct for Theme. It's basically just a repository of constants, anyway. This way, we don't have to deal with any lifecycle issues, or retrieve an instance of it all over the place. Applications can also "extend" it by starting their own Theme struct with ours.